refactor(jina): 계층 구조에 맞게 폴더 구조화 #14
