+++ src/shared/ui/NavBar/Footer/styled.ts
@@ -1,0 +1,33 @@
+import { css } from '@emotion/react';
+import styled from '@emotion/styled';
+
+import IconArrow from 'shared/assets/image/icons/nav-bar/icon-arrow.svg';
+import { CircleButton } from 'shared/ui/NavBar/styled';
+import { NavBarSizeProps } from 'shared/ui/NavBar/type';
+
+export const Footer = styled.div<NavBarSizeProps>`
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  padding: 25px 38px 245px 30px;
+
+  ${({ size }) =>
+    size === 'small' &&
+    css`
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      justify-content: space-between;
+      gap: 164px;
+      padding-bottom: 246px;
+    `};
+`;
+
+export const AddButton = styled(CircleButton)``;
+
+export const FoldButton = styled(CircleButton)`
+  background-image: url(${IconArrow});
+  background-repeat: no-repeat;
+  background-position: center;
+  border: none;
+`;
