+++ src/shared/ui/NavBar/PlayList/PlayListItem/styled.ts
@@ -1,0 +1,11 @@
+import styled from '@emotion/styled';
+
+import { CoverWrapper } from 'shared/ui/NavBar/styled';
+
+export const PlayListInfo = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 30px;
+`;
+
+export const PlayListCoverWrapper = styled(CoverWrapper)``;
