+++ src/shared/ui/NavBar/PlayList/TrackList/TrackItem/index.tsx
@@ -1,6 +1,7 @@
-import TempCoverSrc from '../cover.png';
-import { Cover, CoverWrapper, MoreButton, Title } from '../styled';
-import { NavBarSizeProps } from '../type';
+import TempCoverSrc from 'shared/ui/NavBar/cover.png';
+import { Cover, CoverWrapper, MoreButton, Title } from 'shared/ui/NavBar/styled';
+import { NavBarSizeProps } from 'shared/ui/NavBar/type';
+
 import { ArtistName, TrackInfo, TrackInfoContainer } from './styled';
 
 export const TrackItem = ({ size }: NavBarSizeProps) => {
